refactor: Changed file names, updated imports to new format + add status field to mocked consent in tests Refactored domain/consents/consents.ts to domain/consents/{ID}
